format comments for readability
I added a space after the #. It makes comments easier to pick up and read. You might want to consider changing the content of some of the comments. This isn't a walkthrough, so the "we" language isn't really necessary, along with the statistical reasoning behind things. If a walkthrough is desired, then that should probably be a markdown document or an .Rmd